Make UOM changes happen during setDriver.
The cloud based Polyglot code modifies the node's driver object
internally from an array of dicts to a dict of dicts. This means
that the node's driver object structure is different when running on
the cloud vs. a local Polyglot and we can't easily modify the nodes
driver UOM's at runtime.

To work-a-round that, create a new object, uoms that holds the
runtime uom value for each driver. Then when calling the setDriver
function to set the driver value, also set the driver UOM based on
the runtime value stored in the uoms object.